* Added instrumentation and created compiler check for function signatures * Logging OpenGL messages * Stacklayer iterator fix * Moved from vs2019 -> vs2022 * Small fixes and changes
