+++ Typhoon/src/Platform/OpenGL/OpenGLContext.cpp
@@ -1,5 +1,5 @@
 #include "typhpch.h"
-#include "OpenGLContext.h"
+#include "Platform/OpenGL/OpenGLContext.h"
 
 #include <GLFW/glfw3.h>
 #include <glad/glad.h>
@@ -14,6 +14,8 @@ namespace Typhoon {
 
 	void OpenGLContext::Init()
 	{
+		TYPH_PROFILE_FUNCTION();
+
 		glfwMakeContextCurrent(m_WindowHandle);
 		int status = gladLoadGLLoader((GLADloadproc)glfwGetProcAddress);
 		TYPH_CORE_ASSERT(status, "Failed to initialize Glad!");
@@ -33,6 +35,8 @@ namespace Typhoon {
 
 	void OpenGLContext::SwapBuffers()
 	{
+		TYPH_PROFILE_FUNCTION();
+
 		glfwSwapBuffers(m_WindowHandle);
 	}
