+++ Typhoon/src/Platform/Windows/WindowsInput.cpp
@@ -1,5 +1,5 @@
 #include "typhpch.h"
-#include "WindowsInput.h"
+#include "Platform/Windows/WindowsInput.h"
 
 #include "Typhoon/Core/Application.h"
 #include <GLFW/glfw3.h>
