+++ Typhoon/src/Platform/Windows/WindowsWindow.cpp
@@ -1,9 +1,11 @@
 #include "typhpch.h"
-#include "WindowsWindow.h"
+#include "Platform/Windows/WindowsWindow.h"
 
 #include "Typhoon/Events/ApplicationEvent.h"
 #include "Typhoon/Events/MouseEvent.h"
 #include "Typhoon/Events/KeyEvent.h"
+
+#include "Typhoon/Renderer/Renderer.h"
 
 #include "Platform/OpenGL/OpenGLContext.h"
 
@@ -23,22 +25,30 @@ namespace Typhoon {
 
 	WindowsWindow::WindowsWindow(const WindowProps& props)
 	{
+		TYPH_PROFILE_FUNCTION();
+
 		Init(props);
 	}
 
 	WindowsWindow::~WindowsWindow()
 	{
+		TYPH_PROFILE_FUNCTION();
+
 		Shutdown();
 	}
 
 	void WindowsWindow::OnUpdate()
 	{
+		TYPH_PROFILE_FUNCTION();
+
 		glfwPollEvents();
 		m_Context->SwapBuffers();
 	}
 
 	void WindowsWindow::SetVSync(bool enabled)
 	{
+		TYPH_PROFILE_FUNCTION();
+
 		if (enabled)
 			glfwSwapInterval(1);
 		else
@@ -54,6 +64,8 @@ namespace Typhoon {
 
 	void WindowsWindow::Init(const WindowProps& props)
 	{
+		TYPH_PROFILE_FUNCTION();
+
 		m_Data.Title = props.Title;
 		m_Data.Width = props.Width;
 		m_Data.Height = props.Height;
@@ -63,14 +75,26 @@ namespace Typhoon {
 
 		if (s_GLFWWindowCount == 0)
 		{
-			TYPH_CORE_INFO("Initializing GLFW");
-			int succes = glfwInit();
-			TYPH_CORE_ASSERT(succes, "Could not initialize GLFW");
+			{
+				TYPH_PROFILE_SCOPE("glfw init");
+
+				int succes = glfwInit();
+				TYPH_CORE_ASSERT(succes, "Could not initialize GLFW");
+			}
 			glfwSetErrorCallback(GLFWErrorCallback);
 		}
 
-		m_Window = glfwCreateWindow((int)props.Width, (int)props.Height, m_Data.Title.c_str(), nullptr, nullptr);
-		++s_GLFWWindowCount;
+		{
+			TYPH_PROFILE_SCOPE("glfw Create Window");
+
+			#if defined(TYPH_DEBUG)
+				if (Renderer::GetAPI() == RendererAPI::API::OpenGL)
+					glfwWindowHint(GLFW_OPENGL_DEBUG_CONTEXT, GLFW_TRUE);
+			#endif
+
+			m_Window = glfwCreateWindow((int)props.Width, (int)props.Height, m_Data.Title.c_str(), nullptr, nullptr);
+			++s_GLFWWindowCount;
+		}
 
 		m_Context = GraphicsContext::Create(m_Window);
 		
@@ -174,6 +198,8 @@ namespace Typhoon {
 
 	void WindowsWindow::Shutdown()
 	{
+		TYPH_PROFILE_FUNCTION();
+
 		glfwDestroyWindow(m_Window);
 		--s_GLFWWindowCount;
