+++ Typhoon/src/Typhoon/Core/Log.cpp
@@ -1,5 +1,5 @@
 #include "typhpch.h"
-#include "Log.h"
+#include "Typhoon/Core/Log.h"
 #include <spdlog/sinks/stdout_color_sinks.h>
 
 namespace Typhoon {
