+++ Typhoon/src/Typhoon/Core/Log.h
@@ -1,8 +1,8 @@
 #pragma once
 
 #include "Typhoon/Core/Base.h"
-#include "spdlog/spdlog.h"
-#include "spdlog/fmt/ostr.h"
+#include <spdlog/spdlog.h>
+#include <spdlog/fmt/ostr.h> 
 
 namespace Typhoon {
