+++ Typhoon/src/Typhoon/Debug/Instrumentor.h
@@ -13,6 +13,7 @@
 // You will probably want to macro-fy this, to switch on/off easily and use things like __FUNCSIG__ for the profile name.
 //
 #pragma once
+#include "Typhoon/Core/Base.h"
 
 #include <string>
 #include <chrono>
@@ -21,15 +22,13 @@
 
 #include <thread>
 
-#include "Typhoon/Core/Base.h"
-
 namespace Typhoon {
 
     struct ProfileResult
     {
         std::string Name;
         long long Start, End;
-        uint32_t ThreadID;
+        size_t ThreadID;
     };
 
     class Instrumentor
@@ -127,7 +126,7 @@ namespace Typhoon {
             long long start = std::chrono::time_point_cast<std::chrono::microseconds>(m_StartTimepoint).time_since_epoch().count();
             long long end = std::chrono::time_point_cast<std::chrono::microseconds>(endTimepoint).time_since_epoch().count();
 
-            uint32_t threadID = std::hash<std::thread::id>{}(std::this_thread::get_id());
+            size_t threadID = std::hash<std::thread::id>{}(std::this_thread::get_id());
             Instrumentor::Get().WriteProfile({ m_Name, start, end, threadID });
 
             m_Stopped = true;
@@ -141,10 +140,30 @@ namespace Typhoon {
 }
 
 #if defined(TYPH_ENABLE_PROFILE)
+    #if defined(__GNUC__) || (defined(__MWERKS__) && (__MWERKS__ >= 0x3000)) || (defined(__ICC) && (__ICC >= 600)) || defined(__ghs__)
+        #define TYPH_FUNC_SIG __PRETTY_FUNCTION__
+    #elif defined(__DMC__) && (__DMC__ >= 0x810)
+        #define TYPH_FUNC_SIG __PRETTY_FUNCTION__
+    #elif defined(__FUNCSIG__) // In VS2022 it evaluates to this
+        #define TYPH_FUNC_SIG __FUNCSIG__
+    #elif defined(_MSC_VER) && (_MSC_VER >= 1910)
+        #define TYPH_FUNC_SIG __FUNCTION__
+    #elif (defined(__INTEL_COMPILER) && (__INTEL_COMPILER >= 600)) || (defined(__IBMCPP__) && (__IBMCPP__ >= 500))
+        #define TYPH_FUNC_SIG __FUNCTION__
+    #elif defined(__BORLANDC__) && (__BORLANDC__ >= 0x550)
+        #define TYPH_FUNC_SIG __FUNC__
+    #elif defined(__STDC_VERSION__) && (__STDC_VERSION__ >= 199901)
+        #define TYPH_FUNC_SIG __func__
+    #elif defined(__cplusplus) && (__cplusplus >= 201103)
+        #define TYPH_FUNC_SIG __func__
+    #else
+        #define TYPH_FUNC_SIG "HZ_FUNC_SIG unknown!"
+    #endif
+
     #define TYPH_PROFILE_BEGIN_SESSION(name, filepath) ::Typhoon::Instrumentor::Get().BeginSession(name, filepath)
     #define TYPH_PROFILE_END_SESSION() ::Typhoon::Instrumentor::Get().EndSession()
-    #define TYPH_PROFILE_SCOPE(name) ::Typhoon::InstrumentationTimer timer##__LINE__(name)
-    #define TYPH_PROFILE_FUNCTION() TYPH_PROFILE_SCOPE(__FUNCSIG__)
+    #define TYPH_PROFILE_SCOPE(name) ::Typhoon::InstrumentationTimer TOKENPASTE2(timer, __LINE__)(name)
+    #define TYPH_PROFILE_FUNCTION() TYPH_PROFILE_SCOPE(TYPH_FUNC_SIG)
 #else
     #define TYPH_PROFILE_BEGIN_SESSION(name, filepath)
     #define TYPH_PROFILE_END_SESSION()
