+++ Typhoon/src/Typhoon/ImGui/ImGuiLayer.cpp
@@ -1,12 +1,13 @@
 #include "typhpch.h"
-#include "ImGuiLayer.h"
+#include "Typhoon/ImGui/ImGuiLayer.h"
+
+#include "Typhoon/Core/Application.h"
 
 #include <imgui.h>
 
 #include <backends/imgui_impl_glfw.h>
 #include <backends/imgui_impl_opengl3.h>
 
-#include "Typhoon/Core/Application.h"
 
 // TEMPORARY
 #include <GLFW/glfw3.h>
@@ -26,6 +27,8 @@ namespace Typhoon {
 
 	void ImGuiLayer::OnAttach()
 	{
+		TYPH_PROFILE_FUNCTION();
+
 		// Setup Dear ImGui context
 		IMGUI_CHECKVERSION();
 		ImGui::CreateContext();
@@ -59,6 +62,8 @@ namespace Typhoon {
 
 	void ImGuiLayer::OnDetach()
 	{
+		TYPH_PROFILE_FUNCTION();
+
 		ImGui_ImplOpenGL3_Shutdown();
 		ImGui_ImplGlfw_Shutdown();
 		ImGui::DestroyContext();
@@ -72,6 +77,8 @@ namespace Typhoon {
 
 	void ImGuiLayer::Begin()
 	{
+		TYPH_PROFILE_FUNCTION();
+
 		ImGui_ImplOpenGL3_NewFrame();
 		ImGui_ImplGlfw_NewFrame();
 		ImGui::NewFrame();
@@ -79,6 +86,8 @@ namespace Typhoon {
 
 	void ImGuiLayer::End()
 	{
+		TYPH_PROFILE_FUNCTION();
+
 		ImGuiIO& io = ImGui::GetIO();
 		Application& app = Application::Get();
 		io.DisplaySize = ImVec2((float)app.GetWindow().GetWidth(), (float)app.GetWindow().GetHeight());
