+++ Typhoon/src/Typhoon/Renderer/Renderer.cpp
@@ -1,14 +1,16 @@
 #include "typhpch.h"
-#include "Renderer.h"
+#include "Typhoon/Renderer/Renderer.h"
 #include "Typhoon/Renderer/Renderer2D.h"
 
 
 namespace Typhoon {
 
-	Scope<Renderer::SceneData> Renderer::m_SceneData = CreateScope<Renderer::SceneData>();
+	Scope<Renderer::SceneData> Renderer::s_SceneData = CreateScope<Renderer::SceneData>();
 
 	void Renderer::Init()
 	{
+		TYPH_PROFILE_FUNCTION();
+
 		RenderCommand::Init();
 		Renderer2D::Init();
 	}
@@ -25,7 +27,7 @@ namespace Typhoon {
 
 	void Renderer::BeginScene(OrthographicCamera& camera)
 	{
-		m_SceneData->ViewProjectionMatrix = camera.GetViewProjectionMatrix();
+		s_SceneData->ViewProjectionMatrix = camera.GetViewProjectionMatrix();
 	}
 
 	void Renderer::EndScene()
@@ -37,7 +39,7 @@ namespace Typhoon {
 		const glm::mat4& transform)
 	{
 		shader->Bind();
-		shader->SetMat4("u_ViewProjection", m_SceneData->ViewProjectionMatrix);
+		shader->SetMat4("u_ViewProjection", s_SceneData->ViewProjectionMatrix);
 		shader->SetMat4("u_Transform", transform);
 
 		vertexArray->Bind();
