+++ Typhoon/src/Typhoon/Renderer/Renderer2D.cpp
@@ -1,11 +1,10 @@
 #include "typhpch.h"
-#include "Renderer2D.h"
+#include "Typhoon/Renderer/Renderer2D.h"
 
 #include "Typhoon/Renderer/RenderCommand.h"
 #include "Typhoon/Renderer/Resources/Array.h"
 #include "Typhoon/Renderer/Resources/Shader.h"
 
-#include <glm/gtc/type_ptr.hpp>
 #include <glm/gtc/matrix_transform.hpp>
 
 namespace Typhoon
@@ -21,6 +20,8 @@ namespace Typhoon
 
 	void Renderer2D::Init()
 	{
+		TYPH_PROFILE_FUNCTION();
+
 		s_Data = new Renderer2DStorage();
 		s_Data->QuadVertexArray = VertexArray::Create();
 
@@ -56,17 +57,22 @@ namespace Typhoon
 
 	void Renderer2D::Shutdown()
 	{
+		TYPH_PROFILE_FUNCTION();
+
 		delete s_Data;
 	}
 
 	void Renderer2D::BeginScene(const OrthographicCamera& camera)
 	{
+		TYPH_PROFILE_FUNCTION();
+
 		s_Data->TextureShader->Bind();
 		s_Data->TextureShader->SetMat4("u_ViewProjection", camera.GetViewProjectionMatrix());
 	}
 
 	void Renderer2D::EndScene()
 	{
+		TYPH_PROFILE_FUNCTION();
 
 	}
 
@@ -77,6 +83,8 @@ namespace Typhoon
 
 	void Renderer2D::DrawQuad(const glm::vec3& position, const glm::vec2& size, const glm::vec4& color, const float angle /*= 0.f*/)
 	{
+		TYPH_PROFILE_FUNCTION();
+
 		glm::mat4 scale = glm::scale(glm::mat4(1.f), {size.x, size.y, 1.f});
 		glm::mat4 rotation = glm::rotate(glm::mat4(1.f), glm::radians(angle), { 0.f, 0.f, 1.f });
 		glm::mat4 transform = glm::translate(glm::mat4(1.f), position) * rotation * scale;
@@ -97,6 +105,8 @@ namespace Typhoon
 
 	void Renderer2D::DrawQuad(const glm::vec3& position, const glm::vec2& size, const Ref<Texture2D>& texture, const float angle /*= 0.f*/)
 	{
+		TYPH_PROFILE_FUNCTION();
+
 		s_Data->TextureShader->SetFloat4("u_Color", glm::vec4(1.f));
 		texture->Bind();
