+++ premake5.lua
@@ -1,5 +1,5 @@
 workspace "Typhoon"
-	architecture "x64"
+	architecture "x86_64"
 	startproject "Sandbox"
 
 	configurations
