+++ scripts/Win-GenProjects.bat
@@ -1,5 +1,5 @@
 @echo off
 pushd %~dp0\..\
-call vendor\bin\premake\premake5.exe vs2019
+call vendor\bin\premake\premake5.exe vs2022
 popd
 PAUSE
